+++ src/main/java/roomescape/controller/api/AuthController.java
@@ -1,0 +1,4 @@
+package roomescape.controller.api;
+
+public class AuthController {
+}
