+++ src/main/java/roomescape/controller/page/AuthPageController.java
@@ -1,0 +1,4 @@
+package roomescape.controller.page;
+
+public class AuthPageController {
+}
